Fix in language, images and landing
